paygate_project/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
python-dotenv>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
PyJWT>=2.8,<3.0
sqlparse>=0.5,<0.6
asgiref>=3.8,<4.0
typing_extensions>=4.0,<5.0
django-ratelimit==3.0.1

